14 - SuperBuilder & basic ITSurveyTest

+++ src/main/java/epamers/surwave/services/OptionService.java
@@ -3,6 +3,7 @@ package epamers.surwave.services;
 import epamers.surwave.entities.Option;
 import epamers.surwave.repos.OptionRepository;
 import java.util.List;
+import java.util.NoSuchElementException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -39,7 +40,10 @@ public class OptionService {
       throw new IllegalArgumentException();
     }
 
-    optionRepository.findById(id).orElseThrow();
+    if (!optionRepository.existsById(id)) {
+      throw new NoSuchElementException();
+    }
+
     option.setId(id);
     optionRepository.save(option);
   }
@@ -47,7 +51,10 @@ public class OptionService {
   @Transactional
   public void delete(Long id) {
 
-    optionRepository.findById(id).orElseThrow();
+    if (!optionRepository.existsById(id)) {
+      throw new NoSuchElementException();
+    }
+
     optionRepository.deleteById(id);
   }
 }
