+++ src/main/java/epamers/surwave/services/SurveyService.java
@@ -1,8 +1,13 @@
 package epamers.surwave.services;
 
+import epamers.surwave.entities.Option;
 import epamers.surwave.entities.Survey;
+import epamers.surwave.entities.SurveyState;
 import epamers.surwave.repos.SurveyRepository;
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Set;
+import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -12,6 +17,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class SurveyService {
 
   private final SurveyRepository surveyRepository;
+  private final OptionService optionService;
 
   public List<Survey> getAll() {
 
@@ -29,6 +35,9 @@ public class SurveyService {
     if (survey == null) {
       throw new IllegalArgumentException();
     }
+
+    processSurvey(survey);
+    survey.setState(SurveyState.CREATED);
     return surveyRepository.save(survey);
   }
 
@@ -39,15 +48,33 @@ public class SurveyService {
       throw new IllegalArgumentException();
     }
 
-    surveyRepository.findById(id).orElseThrow();
+
+
+    if (!surveyRepository.existsById(id)) {
+      throw new NoSuchElementException();
+    }
+
     survey.setId(id);
+    processSurvey(survey);
     surveyRepository.save(survey);
   }
 
   @Transactional
   public void delete(Long id) {
 
-    surveyRepository.findById(id).orElseThrow();
+    if (!surveyRepository.existsById(id)) {
+      throw new NoSuchElementException();
+    }
     surveyRepository.deleteById(id);
+  }
+
+  public void processSurvey(Survey survey) {
+
+    Set<Option> options = survey.getOptions().stream()
+        .map(Option::getId)
+        .map(optionService::getById)
+        .collect(Collectors.toSet());
+
+    survey.setOptions(options);
   }
 }
